+++ job_runner/resume.sh
@@ -101,6 +101,7 @@ while read RESUMING; do
 
     JOB_GROUP="${GROUP}"
     JOB_NAME=$(build_job_name "$JOB_GROUP" "$TAG" "train")
+    require_job_name_length "$JOB_NAME"
 
     # Get experiment dataset
     if [[ -z $EXISTING_RESULTS_DATASET ]]; then